+++ src/OpenLoco/src/Ui/Windows/Vehicle.cpp
@@ -1379,9 +1379,9 @@ namespace OpenLoco::Ui::Windows::Vehicle
         }
 
         // 0x4B38FA
-        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] uint32_t* const scrollWidth, uint32_t* const scrollHeight)
+        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] int32_t* const scrollWidth, int32_t* const scrollHeight)
         {
-            *scrollHeight = static_cast<uint32_t>(Common::getNumCars(self) * self.rowHeight);
+            *scrollHeight = Common::getNumCars(self) * self.rowHeight;
         }
 
         // 0x004B3B54
@@ -2467,9 +2467,9 @@ namespace OpenLoco::Ui::Windows::Vehicle
         }
 
         // 0x004B4360
-        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] uint32_t* const scrollWidth, uint32_t* const scrollHeight)
+        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] int32_t* const scrollWidth, int32_t* const scrollHeight)
         {
-            *scrollHeight = static_cast<uint32_t>(Common::getNumCars(self) * self.rowHeight);
+            *scrollHeight = Common::getNumCars(self) * self.rowHeight;
         }
 
         static char* generateCargoTooltipDetails(char* buffer, const StringId cargoFormat, const uint8_t cargoType, const uint8_t maxCargo, const uint32_t acceptedCargoTypes)
@@ -3561,7 +3561,7 @@ namespace OpenLoco::Ui::Windows::Vehicle
         }
 
         // 0x004B4D9B
-        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] uint32_t* const scrollWidth, uint32_t* const scrollHeight)
+        static void getScrollSize(Ui::Window& self, [[maybe_unused]] const uint32_t scrollIndex, [[maybe_unused]] int32_t* const scrollWidth, int32_t* const scrollHeight)
         {
             auto head = Common::getVehicle(self);
             if (head == nullptr)
